feat: allow wallet connetion on single click
