+++ packages/react-kit/package.json
@@ -94,6 +94,7 @@
     "babel-plugin-named-exports-order": "^0.0.2",
     "eslint": "^8.10.0",
     "prop-types": "^15.8.1",
+    "react": "^18.2.0",
     "rimraf": "^3.0.2",
     "storybook-addon-styled-component-theme": "^2.0.0",
     "typescript": "^4.7.4",
