+++ packages/react-kit/src/components/buttons/connectWallet.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import getWeb3Provider from "../../lib/getWeb3Provider";
+import Button from "./button";
+
+const ConnectWallet = () => {
+  return <Button onClick={getWeb3Provider}>Connect to Wallet</Button>;
+};
+
+export default ConnectWallet;
