+++ packages/react-kit/src/stories/buttons/CommitButton.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import CommitButton from "../../components/button/CommitButton";
+import CommitButton from "../../components/buttons/commitButton";
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
