+++ packages/react-kit/src/stories/buttons/button.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import Button from "../../components/button/button";
+import Button from "../../components/buttons/button";
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
